refactor: reusable helper for app config
